+++ src/library/adash-errors.adb
@@ -64,6 +64,8 @@ package body Adash.Errors is
             | Error_No_Such_Slice
             | Error_Needs_Bounds
             | Error_Open_By_Element
+            | Error_Too_Many_At_Once
+            | Error_Too_Many_Parameters
             | Error_Case_Not_Discrete
             | Error_Case_Choice_Not_Static
             | Error_Case_Choice_Covered_Twice
@@ -223,6 +225,10 @@ package body Adash.Errors is
             return Adash.Messages.Msg_Needs_Bounds;
          when Error_Open_By_Element        =>
             return Adash.Messages.Msg_Open_By_Element;
+         when Error_Too_Many_At_Once       =>
+            return Adash.Messages.Msg_Too_Many_At_Once;
+         when Error_Too_Many_Parameters    =>
+            return Adash.Messages.Msg_Too_Many_Parameters;
          when Error_Case_Not_Discrete      =>
             return Adash.Messages.Msg_Case_Not_Discrete;
          when Error_Case_Choice_Not_Static =>